Move CTA to sidebar for better visibility
